test direct-build-windows test direct-build-android, fixes#356#363
