+++ packages/flipper_models/pubspec.yaml
@@ -44,7 +44,7 @@ dependencies:
   talker: ^4.1.5
   talker_flutter: ^4.1.5
   rxdart: ^0.28.0
-  collection: ^1.18.0
+  collection: ^1.19.0
   intl: ^0.19.0
   amplify_core: ^2.5.0
   amplify_api: ^2.5.0
